+++ src/main/java/nextstep/session/service/CoverServiceImpl.java
@@ -3,11 +3,8 @@ package nextstep.session.service;
 import nextstep.common.domain.BaseEntity;
 import nextstep.common.domain.DeleteHistory;
 import nextstep.exception.CoverException;
-import nextstep.session.domain.Cover;
-import nextstep.session.domain.ImageFilePath;
-import nextstep.session.domain.Resolution;
+import nextstep.session.domain.*;
 import nextstep.session.dto.CoverDto;
-import nextstep.session.infrastructure.CoverRepository;
 import nextstep.users.domain.NsUser;
 import nextstep.users.infrastructure.UserService;
 import org.springframework.stereotype.Service;
