+++ src/main/java/nextstep/session/service/StudentServiceImpl.java
@@ -5,9 +5,10 @@ import nextstep.common.domain.DeleteHistory;
 import nextstep.common.domain.DeleteHistoryTargets;
 import nextstep.exception.StudentsException;
 import nextstep.session.domain.Student;
+import nextstep.session.domain.StudentService;
 import nextstep.session.domain.Students;
 import nextstep.session.dto.StudentDto;
-import nextstep.session.infrastructure.StudentRepository;
+import nextstep.session.domain.StudentRepository;
 import nextstep.users.domain.NsUser;
 import nextstep.users.infrastructure.UserService;
 import org.springframework.stereotype.Service;
